refactor of config_loader
