+++ mslib/msui/flighttrack.py
@@ -46,7 +46,8 @@ from PyQt5 import QtGui, QtCore, QtWidgets
 
 from mslib import utils, __version__
 from mslib.utils.units import units
-from mslib.utils import thermolib, config_loader, find_location, save_settings_qsettings, load_settings_qsettings
+from mslib.utils import thermolib, find_location, save_settings_qsettings, load_settings_qsettings
+from mslib.utils.config import config_loader
 from mslib.msui import MissionSupportSystemDefaultConfig as mss_default
 from mslib.msui.mss_qt import variant_to_string, variant_to_float
 from mslib.msui.performance_settings import DEFAULT_PERFORMANCE
