+++ mslib/msui/mscolab_version_history.py
@@ -34,7 +34,8 @@ from werkzeug.urls import url_encode, url_join
 from mslib.msui.flighttrack import WaypointsTableModel
 from PyQt5 import QtCore, QtWidgets, QtGui
 from mslib.msui.qt5 import ui_mscolab_version_history as ui
-from mslib.utils import config_loader, show_popup, utc_to_local_datetime
+from mslib.utils import show_popup, utc_to_local_datetime
+from mslib.utils.config import config_loader
 
 
 class MSColabVersionHistory(QtWidgets.QMainWindow, ui.Ui_MscolabVersionHistory):
